Extract RepreState enum
